member에 member3 덮어씌우기
