+++ hotfist/src/main/java/com/seok/hotfist/repository/MemberRepository.java
@@ -1,7 +1,6 @@
 package com.seok.hotfist.repository;
 
 import com.seok.hotfist.aggregate.Member;
-import com.seok.hotfist.aggregate.Member3;
 import com.seok.hotfist.aggregate.Status;
 import com.seok.hotfist.stream.MyObjectOutput;
 
@@ -12,7 +11,7 @@ public class MemberRepository {
 // 입출력 스트림 (통로)
 // 역할 : DB와 입출력(CRUD) , 성공데이터 또는 실패여부를 Service에게 반환한다. 판단은 Service가 한다
     // 초기 Member에 더미데이값 넣기.
-    private final ArrayList<Member3> memberList = new ArrayList<>();
+    private final ArrayList<Member> memberList = new ArrayList<>();
     private final File file = new File(
             "src/main/java/com/seok/hotfist/db/memberDB.dat");
 
@@ -20,14 +19,14 @@ public class MemberRepository {
     public MemberRepository() {
 
         if (!file.exists()) {           // 파일이 없다면 아래 로직이 진행됨
-            ArrayList<Member3> defaultMemberList = new ArrayList<>();
+            ArrayList<Member> defaultMemberList = new ArrayList<>();
             //파일이 존재하지 않을 때 한번 생성
-            defaultMemberList.add(new Member3(1,"id01","pwd01","강철권고윤석",Status.ACTIVE ));
-            defaultMemberList.add(new Member3(2,"id02","pwd02","폭풍신곽우석",Status.ACTIVE));
-            defaultMemberList.add(new Member3(3,"id03","pwd03","맹수혼최혜민",Status.ACTIVE));
-            defaultMemberList.add(new Member3(4,"id04","pwd04","전광축강수지",Status.ACTIVE));
-            defaultMemberList.add(new Member3(5,"id05","pwd05","암흑룡김석희",Status.ACTIVE));
-            defaultMemberList.add(new Member3(6,"id06","pwd06","섬광참김성민",Status.ACTIVE));
+            defaultMemberList.add(new Member(1,"id01","pwd01","강철권고윤석",Status.ACTIVE ));
+            defaultMemberList.add(new Member(2,"id02","pwd02","폭풍신곽우석",Status.ACTIVE));
+            defaultMemberList.add(new Member(3,"id03","pwd03","맹수혼최혜민",Status.ACTIVE));
+            defaultMemberList.add(new Member(4,"id04","pwd04","전광축강수지",Status.ACTIVE));
+            defaultMemberList.add(new Member(5,"id05","pwd05","암흑룡김석희",Status.ACTIVE));
+            defaultMemberList.add(new Member(6,"id06","pwd06","섬광참김성민",Status.ACTIVE));
 
             saveMembers(defaultMemberList);
         }
@@ -43,7 +42,7 @@ public class MemberRepository {
                 )))
         {
             while (true){
-                memberList.add((Member3)ois.readObject());
+                memberList.add((Member)ois.readObject());
             }
 
         } catch(EOFException e) {
@@ -58,7 +57,7 @@ public class MemberRepository {
     }
 
     //ArrayList<Member3>를 받으면 배열로 출력하는 메소드 (덮어씌우는 기능)
-    private void saveMembers(ArrayList<Member3> inputMemberList) {
+    private void saveMembers(ArrayList<Member> inputMemberList) {
         ObjectOutputStream oos = null;
         try {
             oos = new ObjectOutputStream(
@@ -67,8 +66,8 @@ public class MemberRepository {
 
                     )
             );
-            for (Member3 member3 : inputMemberList) {
-                oos.writeObject(member3);
+            for (Member member : inputMemberList) {
+                oos.writeObject(member);
             }
 
         } catch (IOException e) {
@@ -84,11 +83,11 @@ public class MemberRepository {
 
 
     public int selectLastMemberNo() {
-        Member3 lastMember3 = memberList.get(memberList.size() -1);
-        return lastMember3.getMemNo();
+        Member lastMember = memberList.get(memberList.size() -1);
+        return lastMember.getMemNo();
     }
 
-    public int insertMember(Member3 member3) {
+    public int insertMember(Member member) {
         //헤더가 추가되지 않는 ObjectOutputStream 클래스 정의(MyOutputStream
         MyObjectOutput moo = null;
         int result = 0;
@@ -100,10 +99,10 @@ public class MemberRepository {
                     )
             );
             //파일로 신규회원 추가하기
-            moo.writeObject(member3);
+            moo.writeObject(member);
 
             //컬렉션에도 신규회원 추가하기
-            memberList.add(member3);
+            memberList.add(member);
 
             result = 1;
         } catch (IOException e) {
