+++ hotfist/src/main/java/com/seok/hotfist/service/RankService.java
@@ -25,20 +25,20 @@ public class RankService {
 
         int currentRank = 1;
         int sameRankCount = 0;
-        int previousScore = ranks.get(0).getScore();
+//        int previousScore = ranks.get(0).getScore();
 
         for (int i = 0; i < ranks.size(); i++) {
             Member member = ranks.get(i);
 
-            if (i > 0 && previousScore != member.getScore()) {
-                currentRank += sameRankCount;
-                sameRankCount = 0;
-            }
-
-            System.out.println(currentRank + "위 | " + member.getNickname() + " | " + member.getScore() + "점");
-
-            previousScore = member.getScore();
-            sameRankCount++;
+//            if (i > 0 && previousScore != member.getScore()) {
+//                currentRank += sameRankCount;
+//                sameRankCount = 0;
+//            }
+//
+//            System.out.println(currentRank + "위 | " + member.getNickname() + " | " + member.getScore() + "점");
+//
+//            previousScore = member.getScore();
+//            sameRankCount++;
         }
     }
 }
